+++ 3DRadSpace/Engine3DRadSpace/FFI/Input/Mouse.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../Core/Libs.h"
+#include "../../Core/Libs.h"
 #include "ButtonState.h"
 #include "../Math/Point.h"
