+++ 3DRadSpace/Engine3DRadSpace/FFI/Logging/Warning.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../Core/Libs.h"
+#include "../../Core/Libs.h"
 
 typedef struct E3DRSP_Warning
 {
