+++ 3DRadSpace/Engine3DRadSpace/FFI/Objects/ICamera.cpp
@@ -1,6 +1,8 @@
 #include "../../Objects/ICamera.hpp"
 #include "ICamera.h"
 
+using namespace Engine3DRadSpace::Objects;
+
 E3DRSP_Matrix4x4 E3DRSP_ICamera_GetViewMatrix(E3DRSP_ICamera cam)
 {
 	return std::bit_cast<E3DRSP_Matrix4x4>(static_cast<ICamera*>(cam)->GetViewMatrix());
