+++ 3DRadSpace/Engine3DRadSpace/FFI/Objects/IObject.cpp
@@ -3,7 +3,6 @@
 
 using namespace Engine3DRadSpace;
 using namespace Engine3DRadSpace::Graphics;
-using namespace Engine3DRadSpace::Input;
 using namespace Engine3DRadSpace::Math;
 using namespace Engine3DRadSpace::Objects;
