+++ 3DRadSpace/Engine3DRadSpace/Objects/Impl/CMakeLists.txt
@@ -66,7 +66,7 @@ if(MSVC)
 endif()
 
 target_link_libraries(3DRadSpace.Objects.Impl PRIVATE 3DRadSpace.Audio)
-target_link_libraries(3DRadSpace.Objects.Impl PRIVATE 3DRadSpace.Graphics)
+target_link_libraries(3DRadSpace.Objects.Impl PUBLIC 3DRadSpace.Graphics)
 target_link_libraries(3DRadSpace.Objects.Impl PRIVATE 3DRadSpace.Graphics.Rendering)
 target_link_libraries(3DRadSpace.Objects.Impl PRIVATE 3DRadSpace.Content)
 target_link_libraries(3DRadSpace.Objects.Impl PRIVATE 3DRadSpace.Content.Assets)
