+++ 3DRadSpace/Engine3DRadSpace/Projects/CMakeLists.txt
@@ -32,6 +32,7 @@ target_link_libraries(3DRadSpace.Projects PRIVATE 3DRadSpace.Core)
 target_link_libraries(3DRadSpace.Projects PRIVATE 3DRadSpace.Math)
 target_link_libraries(3DRadSpace.Projects PRIVATE 3DRadSpace.Reflection)
 target_link_libraries(3DRadSpace.Projects PRIVATE 3DRadSpace.Objects)
+target_link_libraries(3DRadSpace.Projects PUBLIC 3DRadSpace.Graphics)
 target_link_libraries(3DRadSpace.Projects PRIVATE 3DRadSpace.Objects.Impl)
 target_link_libraries(3DRadSpace.Projects PRIVATE 3DRadSpace.Objects.Gizmos)
 target_include_directories(3DRadSpace.Projects PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../)
