+++ 3DRadSpace/Engine3DRadSpace/Projects/Serialization.cpp
@@ -1,5 +1,4 @@
 #include "Serialization.hpp"
-#include "Serialization.h"
 #include "../Content/AssetID.hpp"
 #include "../Objects/Impl/Objects.hpp"
 #include "../Objects/ObjectList.hpp"
@@ -526,33 +525,3 @@ bool Engine3DRadSpace::Projects::Serializer::SaveProject(ObjectList* lst, Conten
 	file << std::setw(4) << j;
 	return true;
 } 
-
-bool E3DRSP_LoadProject(
-	E3DRSP_ObjectList* lst,
-	E3DRSP_ContentManager* content,
-	const char* projectPath
-)
-{
-	return Engine3DRadSpace::Projects::Serializer::LoadProject(
-		reinterpret_cast<Engine3DRadSpace::Objects::ObjectList*>(lst),
-		reinterpret_cast<Engine3DRadSpace::Content::ContentManager*>(content),
-		std::filesystem::path(projectPath)
-	);
-}
-
-E3DRSP_IObject E3DRSP_LoadObjectFromProject(const char* path, unsigned id)
-{
-	return static_cast<E3DRSP_IObject>(Engine3DRadSpace::Projects::Serializer::LoadObjectFromProject(
-		std::filesystem::path(path),
-		id
-	));
-}
-
-bool E3DRSP_SaveProject(E3DRSP_ObjectList* lst, E3DRSP_ContentManager* content, const char* projectPath)
-{
-	return Engine3DRadSpace::Projects::Serializer::SaveProject(
-		reinterpret_cast<Engine3DRadSpace::Objects::ObjectList*>(lst),
-		reinterpret_cast<Engine3DRadSpace::Content::ContentManager*>(content),
-		std::filesystem::path(projectPath)
-	);
-}
